refactor: vite를 통해 content_scripts가 번들링 되도록 변경
